Integrate conditional equality cases into existing schema tests
